Refactor module imports to absolute paths and rename run_functions to step_functions
- Changed relative imports to absolute imports in model.py and run.py. This enhances module visibility, making it easier to understand the code structure and reducing the risk of import conflicts in larger projects. It helps when adding tests in a separate directory later.
- Renamed run_functions.py to step_functions.py to more accurately describe the functionality of the contained methods, improving code readability and maintainability. It remove the confusion with run.py, which runs the model, and the step_functions, which are used in the model or agent step.
